Fix potential deadlocks when building proxies/registering stubs in callbacks from events
The Observer design pattern has now been applied in a way that avoids deadlocks,
has better support for concurrency and provides more natural semantics for the
client code:
- An internal mutex is no longer locked while calling back the notify functors (the
listeners) registered by the user.
- The non-interference/concurrency relationship between the notification functions
and the subscribe/unsubscribe functions has remained valid.
- A new notification iteration can now begin in parallel (e.g.: on another thread)
with an on-going notification. The notification iterations do not interfere with
each other.
- The functor callbacks registered by the user with calls to the subscribe function
are now deleted by the call to the unsubscribe function (i.e. deterministically from
the user's point of view), as opposed to being deleted at some point in the future
when/if a notification  happens to be broadcast again (i.e. nondeterministically
from the user's point of view).
- Internally, the implementation of the Event class template has been simplified
(e.g.: by only using one mutex and by letting go of the subscribe/unsubscribe
lists) and the code is now easier to understand.